requirements.txt
spacy
speechrecognition
pyttsx3
spotify
en_core_web_sm-3.1.0-py3-none-any.whl
hugchat